Add Intel XPU (Intel Arc GPU) inference support
Centralize device detection in utils/common.py with XPU in the
priority chain (CUDA > XPU > MPS > CPU).  The three CLI entry points
(infer, infer_batch, demo) now import from common instead of defining
their own helpers.  ASR dtype selection in the model accepts XPU
alongside CUDA for fp16.

Adds an "Intel Arc GPU (XPU)" collapsible section to the main README
under "Step 1: Install PyTorch" with driver, wheel, and verification
instructions.

No new files, no dependency changes, no pyproject.toml modifications.
